Finialized main.cpp format

+++ src/problem14.cpp
@@ -1,14 +1,12 @@
 #include "problem14.h"
-#include "toolkit.h"
-#include <iostream>
 
 using ull = unsigned long long;
 
 int Problem14::solution() {
     int maxStart = 1;
-    ull longestChain = 0;
-    for (ull i = 0; i < 1000000; ++i) {
-        ull currLength = 1;
+    int longestChain = 0;
+    for (int i = 0; i < 1000000; ++i) {
+        int currLength = 1;
         ull j = i;
         while (j > 1) {
             ++currLength;
